Preferences Fix and improved logging around RSS
